パケットサイズを格納する変数の型をu8からu16に変更。

+++ Client/AnpanMMO/Source/AnpanMMO/Packet/PacketHeader.h
@@ -15,7 +15,7 @@ public:
 	PacketID GetPacketId(){ return (PacketID) PacketId; }
 
 	// パケットサイズ取得
-	u8 GetPacketSize(){ return PacketSize; }
+	u16 GetPacketSize(){ return PacketSize; }
 
 	// シリアライズ
 	bool Serialize(MemoryStreamInterface *Stream);
@@ -26,7 +26,7 @@ private:
 	u8 PacketId;
 
 	// パケットサイズ
-	u8 PacketSize;
+	u16 PacketSize;
 
 };
